+++ docs/effect設置格式.lua
@@ -1,0 +1,6 @@
+setting = {
+    name,
+    target,
+    value{percent, number},
+    time,
+}